Fix crash in EImageBase extracting color values from QImage
If pixel data is not stored in a vector the code extracts it directly
from a QImage. Formerly, QImage::color(int index) was used which is plain
wrong because it looks in the color table and not in the pixel data.
The correct API method is QImage::pixel(QPoint pos).